refactor(commitizen): extract a local plugin
