+++ src/commitizen/index.js
@@ -1,0 +1,1 @@
+export {default as scaffold} from './scaffolder.js';
